Fixes incompatibility with prototype
Prototype extends the native javascript objects with its
own functiosn, which causes some problems. This is just
a quick fix to ensure compatibility with a project that
currently uses wikidocs together with prototype.

This problem may be fixable by checking the type of of this
For example if (native_method && obj instanceof Array)

The javascript error that occurred was that this.someFunction
(each if I remember correctly) is not a function.