Namespace the show and hide events so they clean up.
The show and hide events, bound via delegation to the editor itself, did not
clean up properly, so there was this ever growing list of hide/show events on
every math element, which likely caused severe performance issues. Adding
a namespace ensures that it is cleaned up when .bubble is off()'ed.