Let requirejs name the popover plugin instead of explicitly naming it.
Explicitly naming the popover plugin appears to trigger a race condition that
results in a namespace conflict. The issue appears to be that there is both
a directory named popover and a module named popover. The browser loads the
modules asynchronously. If the popover module has finished loading by the time
math, image or link loads, nothing bad happens. If the popover module has not
finished loading, it attempts to load it again, finds the popover directory
and attempts to load lib.js in that directory. By not explicitly naming
the popover module, it will be named popover/popover-plugin automatically,
which in turn resolves the namespace issue.

Refs #347.