When a popover is closed, clear insideScope
insideScope remembers the item on which the popover is currently open. This
is used to determine, on selection change, whether a popover needs to be
opened. It needs to be cleared when a popover is closed, otherwise it becomes
impossible to open a popover on the same element immediately afterwards.

Refs #385.