Move mousedown handler to editor-created event.
This fixes #409. The way it was it would add and remove the mousedown
handler multiple times (for each plugin that uses popovers). It seems a race
condition then ensured that it wasn't added at all. With this change, it is
added once when the editor is created and never removed.