Improve placement logic
Previous logic would drop the popover to the bottom if it fell off the top
or the left. The new code only moves it to the right if it drops off the left,
leaving it at the top. In addition, it drops it to the bottom if it will drop
off the top of the editor, thereby not placing it over the toolbar, at least
initially.