Add "flash" feature to toolbar.
This allows a plugin to alert the user by flashing a button. What it really
does is intermittently apply a class to the button, which you can then
style as you prefer. It comes with no default styling. It also publishes
an event that can be handled by other code, implementing other flashing
mechanisms.