prevent link popover from always wrapping
Decided to still let it wrap if close to the edge. That popover still
needs attention re: the icons, but it's better for now. The change to
the .coffee file was made to prevent extra spacing from getting
underlined, but I couldn't figure out a way to comment that section of
the file, so noting it here.